# BackEnd/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TsunamiIT Manager Project API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/forgotPasswordRequest": {
            "post": {
                "tags": [
                    "Forgot password"
                ],
                "summary": "forgot password request  ",
                "description": "this route allow admin to send an email to reset collaborator's password",
                "operationId": "forgotPasswordRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "tsunamitit.tmp@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Forgot password request created successfully"
                    },
                    "404": {
                        "description": "Email not found "
                    },
                    "500": {
                        "description": "Server Error  "
                    }
                }
            }
        },
        "/api/forgetPasswordVerifyToken": {
            "post": {
                "tags": [
                    "Forgot password"
                ],
                "summary": "Verify token   ",
                "description": "this route verify token for forgot password  ",
                "operationId": "forgetPasswordVerifyToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "xxxxx"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token valid"
                    },
                    "400": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token dont exist"
                    },
                    "500": {
                        "description": "Server Error  "
                    }
                }
            }
        },
...
    }
}